Added interface.
