+++ src/src/Network.cpp
@@ -5,21 +5,8 @@
 namespace SimpleAI {
 
 Network::Network(const Matrix& weights, const Activation& foo)
-  : activation    (foo)
-  , mWeightMatrix (weights)
+  : INetwork (weights, foo)
 {}
-
-List Network::exec(const List& inputs) const
-{
-  List layer = inputs;
-  for (const auto& weightLayer : mWeightMatrix)
-  {
-    layer = execLayer(layer, weightLayer);
-    for (auto& output : layer)
-      output = activation(output);
-  }
-  return layer;
-}
 
 List Network::execLayer(const List& inputs, const List& weights) const
 {
